Duplicate the ImageUploader for PersonImageUploader
Unfortunately, inheriting didn't work as expected because the versions
were only created in the quarantined file store and not in asset
manager. I presume this is because the versions are defined in
ImageUploader and that uploader uses the quarantined file store.

It might be possible to remove some of this duplication but at least
this allows me to make some more progress.