Extract logic for displaying attachment file sizes into a presenter
We want to be able to fetch the file size of an attachment from the
Asset Manager API instead of by analysing the file on disk. This means
we may need to handle various failure cases when making the
request. Extracting the logic used to display the file size to the
user into a presenter will make it easier to isolate the error
handling in one place.