Implement AssetManagerStorage#file_size
This is the equivalent method to
AssetManagerAndQuarantinedFileStorage#file_size for returning a
presented version of the size of the asset. In the case of
AssetManagerStorage, the underlying asset is not available on disk so
we fetch the size from the Asset Manager API. As of [1] all assets in
Asset Manager have this attribute set. There are still some other
things that could go wrong with this method though. Additional error
handling will be added in a subsequent commit.

[1] https://github.com/alphagov/asset-manager/pull/483