Only validate attachment file sizes when files are local
The validation `file_is_not_empty` on AttachmentData is present in
order to prevent users from uploading 0 byte attachments. It relies on
calling the `size` method on the file, which when the file is
originally uploaded or is available on disk can be obtained from
`CarrierWave::SanitizedFile` by `stat`ing the file. When an instance of
AttachmentData is updated, the file has already been uploaded to Asset
Manager and the `file` mounted on AttachmentData is an instance of
`AssetManagerAndQuarantinedFileStorage::File`. To be able to remove the
files entirely from disk we want to be able to switch this storage to
`AssetManagerStorage::File`. This would mean fetching the size of the
file from the Asset Manager API.

Although the API can return a file size, doing so will add an
additional network request, and potentially fail when the API is not
available. Rather than add this network request, we can rely on the
`file_is_not_empty` validation that is triggered when the file is
originally added to ensure that files that have already been stored in
Asset Manager do in fact have a non-zero size. By defining a
`zero_size?` method that always returns `false` for
`AssetManagerAndQuarantinedFileStorage` we ensure that this validation
always passes for assets that are stored in Asset Manager.