Test thumbnail generation by asserting file sent to Asset Manager
The AssetManagerAndQuarantinedFileStorage saves files to disk and
sends them to Asset Manager. Before this commit the tests of PDF
thumbnail generation in AttachmentUploaderPDFTest used the file on
disk to perform the assertions. To make it easier to switch to
AssertManagerStorage we instead change the assertions so that they
test the file that is sent to Asset Manager in the
AssetManagerCreateWhitehallAssetWorker.