Add integration test for discarding draft edition with attachments
This demonstrates the current behaviour when a draft edition, which has
never been published and which has an attachment, is discarded.

* `EditionDeleter#fire_transition!` calls
`Attachable#delete_all_attachments` on the edition.

* This in turn marks each attachment as deleted, but it does *not* call
`Attachment#destroy`.

* This means that even though the only `Attachment` referring to the
relevant instance of `AttachmentData` has been marked as deleted, the
`AttachmentData` is not destroyed.

* So although requests for the attachment URL will result in a `404 Not
 Found`, because of the logic in `AttachmentVisibility#edition_ids` and
`AttachmentsController#fail` mentioned in the previous commit, the
corresponding asset in Asset Manager will not be deleted.

* I think the above somewhat inconsistent behaviour was an oversight in
this commit [1] and I plan to rectify it in a subsequent commit.

[1]:
https://github.com/alphagov/whitehall/commit/1832d02d9fcbb45f5eeb8204548597cccbc7b99f