Destroy AttachmentData when parent edition is discarded
This rectifies the inconsistency in behaviour explained in the previous
commits.

Calling `Attachment#destroy` for each attachment on the edition
triggers the `after_destroy` callback on `FileAttachment` which in turn
will delete its `AttachmentData` and its underlying file and Asset
Manager asset if no other attachments are still referring to the
`AttachmentData`.